move Text component to folder, add style to Text
