+++ src/components/customNative/Text/styles.ts
@@ -1,0 +1,5 @@
+import {StyleSheet} from 'react-native';
+
+export const styles = StyleSheet.create({
+  themeFont: {fontFamily: 'GeezaPro'},
+});
